refactor: initial refactor
